+++ pyperformance/benchmarks/__init__.py
@@ -88,6 +88,7 @@ def BM_2to3(python, options):
     return run_perf_script(python, options, "2to3")
 
 
+@python2_only
 def BM_hg_startup(python, options):
     return run_perf_script(python, options, "hg_startup")