+++ pyperformance/venv.py
@@ -199,10 +199,6 @@ def create_environ(inherit_environ):
         if name in os.environ:
             env[name] = os.environ[name]
 
-    # Needed to install Mercurial on Python 3 for hg_startup:
-    # https://www.mercurial-scm.org/wiki/Python3
-    env['HGPYTHON3'] = '1'
-
     return env