+++ doc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 Version 0.9.2
 -------------
 
-* Enable hg_startup and pyflate benchmarks on Python 3.
+* Enable pyflate benchmarks on Python 3.
 
 Version 0.9.1 (2019-07-29)
 --------------------------